+++ services/portfolio-engine-rs/src/correlation/clusters.rs
@@ -1,5 +1,6 @@
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
+use rust_decimal::Decimal;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 pub enum ClusterState {
@@ -25,8 +26,8 @@ pub struct CorrelationCluster {
     pub cluster_type: ClusterType,
     pub state: ClusterState,
     pub constituent_symbols: Vec<String>,
-    pub internal_correlation: f64,
-    pub weight_in_portfolio: f64,
+    pub internal_correlation: Decimal,
+    pub weight_in_portfolio: Decimal,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -35,5 +36,5 @@ pub struct PortfolioCorrelationResult {
     pub hidden_leverage_detected: bool,
     pub duplicate_themes: Vec<String>,
     pub overexposure_warnings: Vec<String>,
-    pub cross_symbol_concentration: HashMap<String, f64>,
+    pub cross_symbol_concentration: HashMap<String, Decimal>,
 }
