+++ services/portfolio-engine-rs/src/correlation/events.rs
@@ -1,4 +1,5 @@
 use serde::{Deserialize, Serialize};
+use rust_decimal::Decimal;
 
 use super::clusters::PortfolioCorrelationResult;
 
@@ -13,9 +14,9 @@ pub enum MarketRegime {
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum CorrelationEvent {
-    PositionChanged { timestamp: time::OffsetDateTime, symbol: String, new_size: f64 },
+    PositionChanged { timestamp: time::OffsetDateTime, symbol: String, new_size: Decimal },
     RegimeChanged { timestamp: time::OffsetDateTime, new_regime: MarketRegime },
-    VolatilitySpikeDetected { timestamp: time::OffsetDateTime, symbol: String, z_score: f64 },
+    VolatilitySpikeDetected { timestamp: time::OffsetDateTime, symbol: String, z_score: Decimal },
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
